+++ .pre-commit-config.yaml
@@ -1,11 +1,24 @@
 repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.1.0
+    hooks:
+      - id: check-merge-conflict
+      - id: debug-statements
   - repo: https://github.com/timothycrosley/isort
     rev: "5.10.1"
     hooks:
       - id: isort
         additional_dependencies: [toml]
-        args: ["--profile", "black"]
   - repo: https://github.com/psf/black
     rev: "22.3.0"
     hooks:
       - id: black
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.9.2
+    hooks:
+      - id: flake8
+  - repo: https://github.com/markdownlint/markdownlint
+    rev: "v0.11.0"
+    hooks:
+      - id: markdownlint
+        args: ["-r", "~MD013,~MD026"]
