+++ README.md
@@ -4,17 +4,17 @@ A platform for sharing trained machine learning models for chemistry.
 
 ## Overview
 
-This repo acts as a template for creating an interactive web app for one trained model. 
+This repo acts as a template for creating an interactive web app for one trained model.
 
-A [Docker image of the containerised app](https://github.com/ImperialCollegeLondon/SupraShare/pkgs/container/suprashare%2Fsuprashare) is automatically generated [using GitHub actions](https://github.com/ImperialCollegeLondon/SupraShare/blob/develop/.github/workflows/ci.yml) when changes are merged into the main branch. 
+A [Docker image of the containerised app](https://github.com/ImperialCollegeLondon/SupraShare/pkgs/container/suprashare%2Fsuprashare) is automatically generated [using GitHub actions](https://github.com/ImperialCollegeLondon/SupraShare/blob/develop/.github/workflows/ci.yml) when changes are merged into the main branch.
 
-The image is added to the webserver at [suprashare.rcs.ic.ac.uk](https://suprashare.rcs.ic.ac.uk) by manually updating (i) the [docker-compose.yml file](https://docs.docker.com/compose/compose-file/) (ii) the [reverse proxy config](https://caddyserver.com/docs/caddyfile) and (iii) the static index page (front page of the web site), all within the virtual machine running the webserver. 
+The image is added to the webserver at [suprashare.rcs.ic.ac.uk](https://suprashare.rcs.ic.ac.uk) by manually updating (i) the [docker-compose.yml file](https://docs.docker.com/compose/compose-file/) (ii) the [reverse proxy config](https://caddyserver.com/docs/caddyfile) and (iii) the static index page (front page of the web site), all within the virtual machine running the webserver.
 
 ## Development
 
 SupraShare is jointly developed by the [Jelfs Research Group](http://www.jelfs-group.org/) and the Imperial College [Research Software Engineering Team](https://www.imperial.ac.uk/admin-services/ict/self-service/research-support/rcs/research-software-engineering/).
 
-Each newly created app based on this template should include: 
+Each newly created app based on this template should include:
 
 - Details of any pulication and archived data relating to the model, including DOIs
 - A detailed specification of the programming environment used to reproduce results
